Create NewHire_BatchFile
Created this batch file to facilitate my job. When I arrived to the organization we had a form where we have CMD commands written on the page, and we had to run each command to get IP and MAC addresses, plus additional commands to run DNS and Group Policy. Then we had to run a separate batch file to add printers to the PC. I created a batch file to do all the steps on one click and save me some time. This was my first file created.