Create RemoveAdminRights
I created this program to facilitate and automate one of my assignments. I get a list every week where users with Administrative rights have to be compared to the authorized list of system admins. Those not on the list must be removed and a form must be updated. This program reads the list of flagged PCs, compares them to the list of allowed admins, removes the local admin rights from unauthorized users, and updates the list for me.